requirements.txt
flask
playwright
requests
python-dotenv
openai
flask_cors
gunicorn==23.0.0


